Use new indices for places and exact match search
