+++ Tests/IsraelHiking.API.Tests/Controllers/SearchControllerTests.cs
@@ -54,8 +54,8 @@ namespace IsraelHiking.API.Tests.Controllers
         [TestMethod]
         public void GetSearchResults_WithPlaceNameThatDoNotExist_ShouldReturnRegularResults()
         {
-            var place = "place";
-            var searchTerm = "searchTerm, " + place;
+            const string place = "place";
+            const string searchTerm = "searchTerm, " + place;
             var featureLocation = new Coordinate(0.5, 0.5);
             var featureInPlace = new Feature(new Point(featureLocation), new AttributesTable
             {
@@ -69,9 +69,9 @@ namespace IsraelHiking.API.Tests.Controllers
             featureInPlace.SetTitles();
             featureInPlace.SetLocation(featureLocation);
             var featuresInsidePlace = new List<Feature> { featureInPlace };
-            _searchRepository.SearchPlaces(place, Languages.ENGLISH).Returns(new List<IFeature>());
-            _searchRepository.Search("searchTerm", Languages.ENGLISH).Returns(new List<IFeature> { featureInPlace });
-            _searchRepository.GetContainers(featureLocation).Returns(new List<IFeature>());
+            _searchRepository.SearchPlaces(searchTerm, Languages.ENGLISH).Returns([]);
+            _searchRepository.Search("searchTerm", Languages.ENGLISH).Returns([featureInPlace]);
+            _searchRepository.GetContainerName(featureLocation, Languages.ENGLISH).Returns(string.Empty);
 
             var results = _controller.GetSearchResults(searchTerm, Languages.ENGLISH).Result.ToList();
 
@@ -83,15 +83,14 @@ namespace IsraelHiking.API.Tests.Controllers
         [TestMethod]
         public void GetSearchResults_WithPlaceName_ShouldSearchOnlyPlacesInThatPlace()
         {
-            var place = "place";
-            var searchTerm = "searchTerm, " + place;
-            var placeFeature = new Feature(new Polygon(new LinearRing(new[]
-            {
+            const string place = "place";
+            const string searchTerm = "searchTerm, " + place;
+            var placeFeature = new Feature(new Polygon(new LinearRing([
                 new Coordinate(0, 0),
                 new Coordinate(0, 1),
                 new Coordinate(2, 0),
                 new Coordinate(0, 0)
-            })), new AttributesTable
+            ])), new AttributesTable
             {
                 {FeatureAttributes.NAME, place},
                 {FeatureAttributes.ID, "place_id" }
@@ -110,10 +109,7 @@ namespace IsraelHiking.API.Tests.Controllers
             featureInPlace.SetTitles();
             featureInPlace.SetLocation(featureLocation);
             var featuresInsidePlace = new List<IFeature> { featureInPlace };
-            _searchRepository.SearchPlaces(place, Languages.ENGLISH).Returns(new List<IFeature> {placeFeature});
-            _searchRepository
-                .SearchByLocation(Arg.Any<Coordinate>(), Arg.Any<Coordinate>(), "searchTerm", Languages.ENGLISH)
-                .Returns(featuresInsidePlace);
+            _searchRepository.SearchPlaces(searchTerm, Languages.ENGLISH).Returns(featuresInsidePlace);
             _searchRepository.GetContainerName(featureLocation, Languages.ENGLISH).Returns(placeFeature.GetTitle(Languages.ENGLISH));
 
             var results = _controller.GetSearchResults(searchTerm, Languages.ENGLISH).Result.ToList();
@@ -126,25 +122,23 @@ namespace IsraelHiking.API.Tests.Controllers
         [TestMethod]
         public void GetSearchResults_GeometryCollection_ShouldNotFail()
         {
-            var place = "place";
-            var searchTerm = "searchTerm, " + place;
-            var placeFeature = new Feature(new Polygon(new LinearRing(new[]
-            {
+            const string place = "place";
+            const string searchTerm = "searchTerm, " + place;
+            var placeFeature = new Feature(new Polygon(new LinearRing([
                 new Coordinate(0, 0),
                 new Coordinate(0, 1),
                 new Coordinate(2, 0),
                 new Coordinate(0, 0)
-            })), new AttributesTable
+            ])), new AttributesTable
             {
                 {FeatureAttributes.NAME, place},
                 {FeatureAttributes.ID, "place_id" }
             });
             placeFeature.SetTitles();
             var featureLocation = new Coordinate(0.5, 0.5);
-            var featureInPlace = new Feature(new GeometryCollection(new Geometry[]
-                {
+            var featureInPlace = new Feature(new GeometryCollection([
                     new Point(featureLocation)
-                }), new AttributesTable
+                ]), new AttributesTable
                 {
                     {FeatureAttributes.NAME, "name"},
                     {FeatureAttributes.POI_CATEGORY, Categories.HISTORIC},
@@ -157,10 +151,7 @@ namespace IsraelHiking.API.Tests.Controllers
             featureInPlace.SetTitles();
             featureInPlace.SetLocation(featureLocation);
             var featuresInsidePlace = new List<IFeature> { featureInPlace };
-            _searchRepository.SearchPlaces(place, Languages.ENGLISH).Returns(new List<IFeature> { placeFeature });
-            _searchRepository
-                .SearchByLocation(Arg.Any<Coordinate>(), Arg.Any<Coordinate>(), "searchTerm", Languages.ENGLISH)
-                .Returns(featuresInsidePlace);
+            _searchRepository.SearchPlaces(searchTerm, Languages.ENGLISH).Returns(featuresInsidePlace);
             _searchRepository.GetContainerName(Arg.Any<Coordinate>(), Arg.Any<string>()).Returns(placeFeature.GetTitle(Languages.ENGLISH));
 
             var results = _controller.GetSearchResults(searchTerm, Languages.ENGLISH).Result.ToList();
@@ -173,26 +164,24 @@ namespace IsraelHiking.API.Tests.Controllers
         [TestMethod]
         public void GetSearchResults_GeometryCollectionNoContainers_ShouldNotFail()
         {
-            var place = "place";
-            var searchTerm = "searchTerm, " + place;
-            var placeFeature = new Feature(new Polygon(new LinearRing(new[]
-            {
+            const string place = "place";
+            const string searchTerm = "searchTerm, " + place;
+            var placeFeature = new Feature(new Polygon(new LinearRing([
                 new Coordinate(0, 0),
                 new Coordinate(0, 1),
                 new Coordinate(2, 0),
                 new Coordinate(0, 0)
-            })), new AttributesTable
+            ])), new AttributesTable
             {
                 {FeatureAttributes.NAME, place},
                 {FeatureAttributes.ID, "place_id" }
             });
             placeFeature.SetTitles();
             var featureLocation = new Coordinate(0.5, 0.5);
-            var featureInPlace = new Feature(new GeometryCollection(new Geometry[]
-                {
+            var featureInPlace = new Feature(new GeometryCollection([
                     new Point(featureLocation),
-                    new LineString(new [] { new Coordinate(0,0), new Coordinate(3,3) })
-                }), new AttributesTable
+                    new LineString([new Coordinate(0,0), new Coordinate(3,3)])
+                ]), new AttributesTable
                 {
                     {FeatureAttributes.NAME, "name"},
                     {FeatureAttributes.POI_CATEGORY, Categories.HISTORIC},
@@ -205,11 +194,8 @@ namespace IsraelHiking.API.Tests.Controllers
             featureInPlace.SetTitles();
             featureInPlace.SetLocation(featureLocation);
             var featuresInsidePlace = new List<IFeature> { featureInPlace };
-            _searchRepository.SearchPlaces(place, Languages.ENGLISH).Returns(new List<IFeature> { placeFeature });
-            _searchRepository
-                .SearchByLocation(Arg.Any<Coordinate>(), Arg.Any<Coordinate>(), "searchTerm", Languages.ENGLISH)
-                .Returns(featuresInsidePlace);
-            _searchRepository.GetContainers(featureLocation).Returns(new List<IFeature> { placeFeature });
+            _searchRepository.SearchPlaces(searchTerm, Languages.ENGLISH).Returns(featuresInsidePlace);
+            _searchRepository.GetContainers(featureLocation).Returns([placeFeature]);
 
             var results = _controller.GetSearchResults(searchTerm, Languages.ENGLISH).Result.ToList();
 
@@ -221,8 +207,8 @@ namespace IsraelHiking.API.Tests.Controllers
         [TestMethod]
         public void GetSearchResults_ContainerHasNoName_ShouldNotIAddItToDisplayName()
         {
-            var place = "place";
-            var searchTerm = "searchTerm";
+            const string place = "place";
+            const string searchTerm = "searchTerm";
             var featureLocation = new Coordinate(0.5, 0.5);
             var featureInPlace = new Feature(new Point(featureLocation), new AttributesTable
             {
@@ -235,7 +221,7 @@ namespace IsraelHiking.API.Tests.Controllers
             });
             featureInPlace.SetTitles();
             featureInPlace.SetLocation(featureLocation);
-            _searchRepository.Search(searchTerm, Languages.ENGLISH).Returns(new List<IFeature> { featureInPlace });
+            _searchRepository.Search(searchTerm, Languages.ENGLISH).Returns([featureInPlace]);
             _searchRepository.GetContainerName(Arg.Any<Coordinate>(), Arg.Any<string>()).Returns(place);
 
             var results = _controller.GetSearchResults(searchTerm, Languages.ENGLISH).Result;
